-Cross-platform(/compiler) UNUSED preprocessor macro which is used to get rid of unused variable/statement with no effect warnings in IMessageHandler.h and INetworkServerListener.h. For lack of a better location, resides in the very beginning of Types.h for now. -Fix signed/unsigned mismatch warning in NetworkSimulator.cpp. -Add missing forward declaration of struct EndPoint in INetworkServerListener.h.

+++ src/NetworkSimulator.cpp
@@ -67,7 +67,7 @@ void NetworkSimulator::SubmitSendBuffer(kNet::OverlappedTransferBuffer *buffer, 
         b.buffer = socket->BeginSend(buffer->bytesContains);
 		if (b.buffer)
 		{
-			assert(b.buffer->buffer.len >= buffer->bytesContains);
+			assert(b.buffer->buffer.len >= (u32)buffer->bytesContains);
 			memcpy(b.buffer->buffer.buf, buffer->buffer.buf, buffer->bytesContains);
 			b.buffer->bytesContains = buffer->bytesContains;
